+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [3.1.0] (IN PROGRESS)
 
+* Provider Results UI updates (UIEH-846)
 * Fix Unsaved Changes Modal not appearing after changing selected status. (UIEH-854)
 * Fix need to click 'X' twice to return after changing selected status and visiting Edit page. (UIEH-854)
 * Fix Proxy selection is set to None when package has Inherited proxy. (UIEH-854)